requirements.txt
# Core API y servidor
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.12
python-dotenv==1.0.0

# Machine Learning y reconocimiento facial
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
mediapipe==0.10.8
opencv-python==4.8.1.78
onnxruntime-gpu==1.17.1   # Si quieres aprovechar tu RTX 4060 (usa onnxruntime==1.17.1 si solo CPU)
scipy==1.11.3             # Necesario para distance.cosine

# Streamlit para la interfaz
streamlit==1.26.0
altair==5.0.1
Pillow<10  # Compatible con Streamlit 1.26.0

# Base de datos
sqlalchemy==2.0.23

# Otros esenciales
requests==2.31.0
matplotlib==3.7.2

# Fijar typing-extensions compatible con TensorFlow
typing-extensions==4.5.0
